* modified the starting position of iteration in AddAssocReactionTerms() (consider revising)   * Added a new class pairedMoleucles   * privatise protected member variables in Molecule.h & setup accessors/mutators for these variables   * rearranged the order initializing a molecule in classes   * Changed temp into beta   * include dMartix and Matrix into namespace mesmer   * removed some typecast warnings in VC compiler   * removed m_alloc   * Add lines in bool MoleculeManager::addmols() preventing user from inputting duplicated molecules   * Modified oberror report level from obWarning to obInfo.   * moved m_SpinMultiplicity to ModelledMolecule   * some other minor modifications
SVN: r80@trunk

+++ src/TimeCounter.cpp
@@ -11,7 +11,7 @@ namespace mesmer
     struct std::tm* timeinfo; timeinfo = std::localtime (&tnow);
     char buffer [20]; std::strftime(buffer,20,"%Y%m%d_%H%M%S", timeinfo);
     std::string myTime(buffer);
-    if (TimeMap.size() > 1) timeElapsed = TimeMap[TimeMap.size()-1].timeStamp - TimeMap[TimeMap.size()-2].timeStamp;
+    if (static_cast<int>(TimeMap.size()) > 1) timeElapsed = TimeMap[TimeMap.size()-1].timeStamp - TimeMap[TimeMap.size()-2].timeStamp;
     else timeElapsed = 0;
     return myTime;
   }
@@ -54,7 +54,7 @@ namespace mesmer
       if (TimeMap[i].stampName == timeStampName){
         tStamp = TimeMap[i].timeStamp;
         foundTimeStamp = true;
-        if (i > 0) timeElapsed = TimeMap[i].timeStamp - TimeMap[i-1].timeStamp;
+        if (i > 0) timeElapsed = static_cast<unsigned int>(TimeMap[i].timeStamp - TimeMap[i-1].timeStamp);
         else timeElapsed = 0;
         break;
       }
